requirements.txt
numpy>=1.18.5
pandas>=1.1.2
matplotlib>=3.2.2
pillow>=7.1.2
scipy>=1.5.1